test(EXUI-4510): move task event data to setup
