+++ playwright_tests_new/integration/test/manageTasks/caseTaskList/taskEventCompletion.setup.ts
@@ -1,4 +1,7 @@
 import type { Page } from '@playwright/test';
+import { buildCaseEventTriggerMock } from '../../../mocks/caseEventTrigger.builder';
+import { buildCaseDetailsTasksFromParams, buildCaseDetailsTasksMinimal } from '../../../mocks/caseDetailsTasks.builder';
+import { buildAsylumCaseMock } from '../../../mocks/cases/asylumCase.mock';
 
 type TaskEventCompletionRouteOptions = {
   assigneeId: string;
@@ -20,6 +23,106 @@ export type TaskEventCompletionSubmissionCapture = {
   }>;
   submittedEvents: Array<Record<string, unknown>>;
 };
+
+const caseId = '1652161372854637';
+const eventId = 'text';
+const eventName = 'Test event';
+const eventFieldId = 'TaskCompletionNote';
+const eventFieldLabel = 'Completion note';
+const taskId = '11111111-2222-4333-8444-555555555555';
+
+const textFieldType = {
+  id: 'Text',
+  type: 'Text',
+  min: null,
+  max: null,
+  regular_expression: null,
+  fixed_list_items: [],
+  complex_fields: [],
+  collection_field_type: null,
+};
+
+export const taskEventCompletionTestData = {
+  userIdentifier: 'STAFF_ADMIN',
+  caseId,
+  eventId,
+  eventName,
+  eventFieldId,
+  eventFieldLabel,
+  taskId,
+  caseMockResponse: buildAsylumCaseMock({
+    caseId,
+    triggers: [{ id: eventId, name: eventName }],
+  }),
+  eventTriggerResponse: {
+    ...buildCaseEventTriggerMock({
+      caseId,
+      eventId,
+      eventName,
+      description: 'Complete the selected case task event',
+    }),
+    case_fields: [
+      {
+        id: eventFieldId,
+        label: eventFieldLabel,
+        hidden: null,
+        order: null,
+        metadata: false,
+        case_type_id: null,
+        hint_text: null,
+        field_type: textFieldType,
+        security_classification: 'PUBLIC',
+        live_from: null,
+        live_until: null,
+        show_condition: null,
+        acls: [],
+        complexACLs: [],
+        display_context: 'OPTIONAL',
+        display_context_parameter: null,
+        retain_hidden_value: null,
+        formatted_value: null,
+        category_id: null,
+      },
+    ],
+    wizard_pages: [
+      {
+        id: `${eventId}Page1`,
+        label: eventName,
+        order: 1,
+        wizard_page_fields: [
+          {
+            case_field_id: eventFieldId,
+            order: 1,
+            page_column_no: null,
+            complex_field_overrides: [],
+          },
+        ],
+        show_condition: null,
+        callback_url_mid_event: null,
+        retries_timeout_mid_event: [],
+      },
+    ],
+  },
+};
+
+export function buildTaskEventCompletionTaskResponses(assigneeId: string) {
+  const visibleTasks = buildCaseDetailsTasksMinimal({
+    caseId,
+    id: [taskId],
+    titles: ['Task 1'],
+    states: ['assigned'],
+    assignees: [assigneeId],
+  });
+  const completableTasksResponse = {
+    task_required_for_event: true,
+    tasks: buildCaseDetailsTasksFromParams({
+      caseId,
+      tasks: [{ id: taskId, title: 'Task 1', state: 'assigned', assignee: assigneeId }],
+    }),
+  };
+
+  return { completableTasksResponse, visibleTasks };
+}
 
 function buildCaseworkerResponse(assigneeId: string) {
   return [
